requirements.txt
PyDrive==1.3.1
discord.py==0.16.12
numpy==1.15.0
discord==0.0.2
GitPython==2.1.11
